requirements.txt
python-telegram-bot[callback-data]>=22.0
httpx>=0.27.0
python-dotenv>=1.0.0
